automation/requirements.txt
pandas
fuzzywuzzy
python-Levenshtein
unidecode
supabase
python-dotenv
